% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmers.R
\name{count_kmers}
\alias{count_kmers}
\title{count_kmers}
\usage{
count_kmers(aa, k = 6, threads = 1, aa2int = FALSE)
}
\arguments{
\item{aa}{character vector/AAString/AAStringSet}

\item{k}{kmer length [default: 6]}

\item{threads}{number of parallel threads [default: 1]}

\item{aa2int}{convert aa to Int64 [default: FALSE]}
}
\value{
\code{table} or \code{list} of \code{table}
}
\description{
This function count kmers from a character vector or an
AAString or an AAStringSet
}
\examples{
data(hiv, package="MSA2dist")
l <- hiv |>
    MSA2dist::cds2aa() |>
    korthoR::count_kmers(k=6)
l
# to convert aa to Int64
lint64 <- hiv |>
    MSA2dist::cds2aa() |>
    korthoR::count_kmers(
        k=6,
        aa2int=TRUE)
lint64
}
\seealso{
\code{\link[Biostrings]{XStringSet-class}}
}
\author{
Kristian K Ullrich
}
