% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmers2.R
\name{count_kmers2}
\alias{count_kmers2}
\title{count_kmers2}
\usage{
count_kmers2(aa, k = 6)
}
\arguments{
\item{aa}{character vector/AAString/AAStringSet}

\item{k}{kmer length [default: 6]}
}
\value{
\code{table} or \code{list} of \code{table}
}
\description{
This function count kmers from a character vector or an
AAString or an AAStringSet
}
\examples{
data(hiv, package="MSA2dist")
l <- hiv |>
    MSA2dist::cds2aa() |>
    korthoR::count_kmers2(k=6)
l
}
\seealso{
\code{\link[Biostrings]{XStringSet-class}}
}
\author{
Kristian K Ullrich
}
