% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_mers2.R
\name{count_mers2}
\alias{count_mers2}
\title{count_mers2}
\usage{
count_mers2(aa, threads = 1)
}
\arguments{
\item{aa}{character vector/AAString/AAStringSet}

\item{threads}{number of parallel threads [default: 1]}
}
\value{
\code{table} or \code{list} of \code{table}
}
\description{
This function count all mers from a character vector or an
AAString or an AAStringSet
}
\examples{
data(hiv, package="MSA2dist")
l <- hiv |>
    MSA2dist::cds2aa() |>
    korthoR::count_mers2()
l
}
\seealso{
\code{\link[Biostrings]{XStringSet-class}}
}
\author{
Kristian K Ullrich
}
