% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bionjs_tree.R
\name{get_bionjs_tree}
\alias{get_bionjs_tree}
\title{get_bionjs_tree}
\usage{
get_bionjs_tree(jaccard_df, value = "jaccard")
}
\arguments{
\item{jaccard_df}{data.frame from \code{get_jaccard_a_b}}

\item{value}{select value from jaccard result to be used as distances to
calculate the bionjs tree [default: jaccard]; choices are "jaccard" or "mash"
or "sumdist"}
}
\value{
\code{data.frame}
}
\description{
This function take the results from \code{get_jaccard_a_b}
and use the distances to calculate a \code{bionjs} tree with the
\code{ape} package
}
\examples{
data(hiv, package="MSA2dist")
l <- hiv |>
    MSA2dist::cds2aa() |>
    korthoR::count_kmers(k=6)
d <- korthoR::get_jaccard_a_b(
    kmer_counts_q=l,
    kmer_counts_t=l,
    k=6)
t1 <- korthoR::get_bionjs_tree(jaccard_df=d, value="jaccard")
t2 <- korthoR::get_bionjs_tree(jaccard_df=d, value="mash")
t3 <- korthoR::get_bionjs_tree(jaccard_df=d, value="sumdist")
par(mfrow=c(1,3))
plot(t1, main="jaccard")
plot(t2, main="mash")
plot(t3, main="sumdist")
}
\author{
Kristian K Ullrich
}
