% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nonzero_from_sparse.R
\name{get_nonzero_from_sparse}
\alias{get_nonzero_from_sparse}
\title{get_nonzero_from_sparse}
\usage{
get_nonzero_from_sparse(mm)
}
\arguments{
\item{mm}{Matrix}
}
\value{
\code{Matrix}
}
\description{
This function count all mers from a character vector
}
\examples{
mm <- Matrix::Matrix(0, ncol=3, nrow=3, sparse = TRUE)
mm[1,1] <- 1
mm |> get_nonzero_from_sparse()
}
\author{
Kristian K Ullrich
}
