% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_count_kmers}
\alias{rcpp_count_kmers}
\title{rcpp_count_kmers}
\arguments{
\item{aavector}{StringVector [mandatory]}

\item{k}{kmer length [default: 6]}

\item{ncores}{number of cores [default: 1]}

\item{aa2int}{convert aa to Int64 [default: FALSE]}
}
\value{
list
}
\description{
returns count kmers
}
\examples{
## load example sequence data
data("hiv", package="MSA2dist")
h <- hiv |> MSA2dist::cds2aa() |> as.character()
l <- korthoR::rcpp_count_kmers(
    aavector=h,
    k=6,
    ncores=1)
l
# to convert aa to Int64
lint64 <- korthoR::rcpp_count_kmers(
    aavector=h,
    k=6,
    ncores=1,
    aa2int=TRUE)
lint64
}
\author{
Kristian K Ullrich
}
