% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_count_mers}
\alias{rcpp_count_mers}
\title{rcpp_count_mers}
\arguments{
\item{aavector}{StringVector [mandatory]}

\item{ncores}{number of cores [default: 1]}
}
\value{
list
}
\description{
returns count all mers
}
\examples{
## load example sequence data
data("hiv", package="MSA2dist")
h <- hiv |> MSA2dist::cds2aa() |> as.character()
l <- korthoR::rcpp_count_mers(
    aavector=h,
    ncores=1)
l
}
\author{
Kristian K Ullrich
}
