% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_jaccard_sparse}
\alias{rcpp_jaccard_sparse}
\title{rcpp_jaccard_sparse}
\arguments{
\item{kmer_counts_q}{list [mandatory]}

\item{kmer_counts_t}{list [mandatory]}

\item{k}{kmer length [default: 6]}

\item{min_jaccard}{min jaccard distance to report [default: 0.01]}

\item{sparse_threshold}{use sparse_n number subset to evaluate change search strategy [default: 0.1]}

\item{sparse_n}{number of subset to get sparse threshold [default: 100]}

\item{ncores}{number of cores [default: 1]}

\item{debug}{print to console [default: FALSE]}
}
\value{
list
}
\description{
returns jaccard
}
\examples{
## load example sequence data
data("hiv", package="MSA2dist")
l <- hiv |>
    MSA2dist::cds2aa() |>
    korthoR::count_kmers(k=6)
d <- korthoR::rcpp_jaccard_sparse(
    kmer_counts_q=l,
    kmer_counts_t=l,
    k=6,
    min_jaccard=0.01,
    sparse_threshold=0.1,
    sparse_n=100,
    ncores=1,
    debug=FALSE)
d
}
\author{
Kristian K Ullrich
}
