% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_jaccard_sparse_a_b_Int64}
\alias{rcpp_jaccard_sparse_a_b_Int64}
\title{rcpp_jaccard_sparse_a_b_Int64}
\arguments{
\item{kmer_counts_q_Int64}{list [mandatory]}

\item{kmer_counts_t_Int64_Int64}{list [mandatory]}

\item{k}{kmer length [default: 6]}

\item{min_jaccard}{min jaccard distance to report [default: 0.01]}

\item{ncores}{number of cores [default: 1]}

\item{debug}{print to console [default: FALSE]}
}
\value{
list
}
\description{
returns jaccard
}
\examples{
## load example sequence data
data("hiv", package="MSA2dist")
lint64 <- hiv |>
    MSA2dist::cds2aa() |>
    korthoR::count_kmers(
        k=6,
        aa2int=TRUE)
d <- korthoR::rcpp_jaccard_sparse_a_b_Int64(
    kmer_counts_q_Int64=lint64,
    kmer_counts_t_Int64=lint64,
    k=6,
    min_jaccard=0.01,
    ncores=1,
    debug=FALSE)
d
}
\author{
Kristian K Ullrich
}
