% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_list_into_chunks.R
\name{split_list_into_chunks}
\alias{split_list_into_chunks}
\title{split_list_into_chunks}
\usage{
split_list_into_chunks(l, num_per_chunk = 1000)
}
\arguments{
\item{l}{list to split}

\item{num_per_chunk}{specify max number per chunk [default: 1000]}
}
\value{
\code{table}
}
\description{
This function splits a list into chunks into at most
num_per_chunk elements
}
\examples{
korthoR::split_list_into_chunks(
    l=vector("list", length=10),
    num_per_chunk=2)
}
\author{
Kristian K Ullrich
}
