% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string2kmers.R
\name{string2kmers}
\alias{string2kmers}
\title{string2kmers}
\usage{
string2kmers(string, k = 6)
}
\arguments{
\item{string}{character vector}

\item{k}{kmer length [default: 6]}
}
\value{
\code{table}
}
\description{
This function count kmers from a character
}
\examples{
korthoR::string2kmers(
    "AACGTGCA",
    k=6)
}
\author{
Kristian K Ullrich
}
