% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string2mers.R
\name{string2mers}
\alias{string2mers}
\title{string2mers}
\usage{
string2mers(string)
}
\arguments{
\item{string}{character vector}
}
\value{
\code{table}
}
\description{
This function count all mers from a character vector
}
\examples{
korthoR::string2mers(
    "AACGTGCA")
}
\author{
Kristian K Ullrich
}
