% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jaccard_a_b.R
\name{get_jaccard_a_b}
\alias{get_jaccard_a_b}
\title{get_jaccard_a_b}
\usage{
get_jaccard_a_b(
  kmer_counts_q,
  kmer_counts_t,
  k = 6,
  min_jaccard = 0.01,
  use_rcpp = TRUE,
  use_sparse = TRUE,
  threads = 1,
  num_per_chunk_q = 100,
  num_per_chunk_t = 1000,
  aa2int = FALSE
)
}
\arguments{
\item{kmer_counts_q}{kmer_counts}

\item{kmer_counts_t}{kmer_counts}

\item{k}{kmer length [default: 6]}

\item{min_jaccard}{min jaccard distance to report [default: 0.01]}

\item{use_rcpp}{use rcpp jaccard distance calculation[ default: TRUE]}

\item{use_sparse}{use rcpp sparse approach [default: TRUE]}

\item{threads}{number of parallel threads [default: 1]}

\item{num_per_chunk_q}{specify max number per chunk q [default: 100]}

\item{num_per_chunk_t}{specify max number per chunk t [default: 1000]}

\item{aa2int}{convert aa to Int64 [default: FALSE]}
}
\value{
\code{data.frame}
}
\description{
This function calculates jaccard distance for all pairwise
comparison between two kmer_counts list objects.
If the parameter k is set, mash distance will be reported.
Only jaccard distances will be reported if threshold is met.
}
\examples{
data(hiv, package="MSA2dist")
l <- hiv |>
    MSA2dist::cds2aa() |>
    korthoR::count_kmers(k=6)
df <- korthoR::get_jaccard_a_b(
    kmer_counts_q=l,
    kmer_counts_t=l,
    k=6,
    min_jaccard=0.01)
df
# to use multiple threads
#df <- korthoR::get_jaccard_a_b(
#    kmer_counts_q=l,
#    kmer_counts_t=l,
#    k=6,
#    min_jaccard=0.01,
#    threads=2)
#df
# to use multiple threads and change chunk size
#df <- korthoR::get_jaccard_a_b(
#    kmer_counts_q=l,
#    kmer_counts_t=l,
#    k=6,
#    min_jaccard=0.01,
#    threads=2,
#    num_per_chunk=5)
#df
# to use Int64 kmers
# lint64 <- hiv |>
#     MSA2dist::cds2aa() |>
#     korthoR::count_kmers(
#     k=6,
#     aa2int=TRUE)
#df <- korthoR::get_jaccard_a_b(
#    kmer_counts_q=lint64,
#    kmer_counts_t=lint64,
#    k=6,
#    min_jaccard=0.01,
#    threads=2,
#    num_per_chunk=5,
#    aa2int=TRUE)
#df
}
\author{
Kristian K Ullrich
}
